All post routes now follow an MVC format
